Update Assignment1_partB.md
Changed from tutorial to Example. In this example you can get the code for a solution of the part B1. Be aware that the plotting does not necessarily match what you can do in your Docker container, if you want to use it as it is, then install the missing dependencies, but the new installations won't be saved after, so you will end up having to install them every time. Better find an alternative that is similar and that is already installed in the Docker container.